cf-backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coloredlogs==15.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
dnspython==2.2.1
drf-yasg==1.21.4
eventlet==0.33.1
flatbuffers==23.3.3
greenlet==1.1.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpath==0.82
MarkupSafe==2.1.1
mpmath==1.3.0
mysqlclient==2.1.1
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==21.3
Pillow==9.2.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.1.0
pyparsing==3.0.9
pyreadline3==3.4.1
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2022.4
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
nltk==3.8.1